🚚 Chore: rename files of utility-service
