+++ apps/web/src/types/utility-service/index.d.ts
@@ -1,3 +1,3 @@
-export * from "./followService";
-export * from "./bookmarkService";
-export * from "./likeService";
+export * from "./follow-service";
+export * from "./bookmark-service";
+export * from "./like-service";
